Add changes to support the  rdf:langString type from RDF 1.1 .
RDF 1.1 gets rid of "plain literals", and instead defines "simple literals".
These have the same lexical form as "plain literals", but are defined as syntactic sugar for either xsd:string or rdf:langString, which partition the value space of PlainLiteral based on the absence or presence of a language tag.

rdf:langString is not used as an explicit datatype in rdf serializations (just like PlainLiteral).   It can appear as the range of a datatype.

The OWL 2 and RDF 1.1 specifications are not currently aligned (e.g. explicit xsd:string  types may be discarded by RDF processors).  The rdf:langRange facet is currently only formally defined for rdf:PlainLiteral, as previously this was the only datatype which could have a language range.

Since rdf:langString can be defined as a PlainLiteral with a langRange restriction of "*", I am allowing this facet to also be used on rdf:langString.