PelletReasoner enabled to work with Fuseki assembler
Constructor adjusted so that schema is indicated as bound when argument is not null.
PelletInfGraph will now retrieve the schema when either bindSchema or bindFixedSchema is used.
Previously bindSchema would have no effect despite this being a Jena Reasoner interface method.
This change enables Pellet to work with Fuseki assembler ja:schema command.

- Added override tags.
- Log set to final.
- Reformatted page for none aligned methods.
- Deprecated bindFixedSchema methods as now repetition of bindSchema and not part of Jena Reasoner interface.
- Added JavaDoc for bindSchema and bindFixedSchema methods to make clearer that the returned Reasoner needs to be used.