First pass on implementation of server model with Protege Server
Decorating some of the immutable states created for the server model to
be able to use them with Protege Server, but we might just declare them
abstract and derivate everything else from there.
We'll probably need some method to refresh the server state to fetch
current ontology changes, maybe there's a hook up for events with the
pellet server otherwise a polling will be required to fetch latest
server snapshots.